\name{US_state_precipitation}
\docType{data}
\alias{US_state_precipitation}
\title{Average state-wide precipitation for US states dating back to 1925.}
\description{
Time-series of average state-wide temperatures (in inches) of the lower 48 US states measured at annual intervals from July 1925 to July 2019. Provided by the National Oceanic and Atmospheric Administration (NOAA).
}
\usage{US_state_precipitation}
\format{A matrix containing 95 observations (rows) from 48 states (columns).}
\source{NOAA <https://www.ncdc.noaa.gov/cag/statewide/time-series>}
